* generic/gdbtk-stack.c (gdb_get_vars_command): Revise cast of 	`clientData' to avoid build error on 64-bit hosts.

+++ gdb/gdbtk/generic/gdbtk-stack.c
@@ -285,7 +285,7 @@ gdb_get_vars_command (ClientData clientData, Tcl_Interp *interp,
       return TCL_ERROR;
     }
 
-  arguments = (int) clientData;
+  arguments = (long) clientData;
 
   /* Initialize the result pointer to an empty list. */
